requirements.txt
Flask==3.0.0
pandas==2.1.3
pymongo==4.6.0
python-dotenv==1.0.0
sendgrid==6.11.0
Werkzeug==3.0.1
urllib3==2.0.7